+++ src/ServiceProvider.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Avtocod\B2BApi\Laravel;
+
+use Illuminate\Support\ServiceProvider as IlluminateServiceProvider;
+
+class ServiceProvider extends IlluminateServiceProvider
+{
+    /**
+     * Register package services.
+     *
+     * @return void
+     */
+    public function register(): void
+    {
+        //
+    }
+
+    /**
+     * Bootstrap package services.
+     *
+     * @return void
+     */
+    public function boot(): void
+    {
+        //
+    }
+}