+++ tests/ServiceProviderTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Avtocod\B2BApiLaravel\Tests;
+
+class ServiceProviderTest extends AbstractTestCase
+{
+    //
+}