+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v3.0.0
+
+### Changed
+
+- Minimal `PHP` version now is `^7.1.3`
+- Maximal `Laravel` version now is `5.8.x`
+- Dependency `avto-dev/b2b-api-php` changed to `avtocod/b2b-api-php`
+- Dependency `laravel/framework` changed to `illuminate/*`
+- Root namespace changed from `\Avtocod\B2BApiLaravel` to `Avtocod\B2BApi\Laravel`
+
+### Removed
+
+- Dependency `avto-dev/b2b-api-php-laravel`
+
 ## v2.2.0
 
 ### Changed